дорабатывает main-page
