+++ src/components/main-page/header-auth.tsx
@@ -1,0 +1,39 @@
+import { Link } from 'react-router-dom';
+import { useAppSelector } from '../hooks';
+import { AppRoute } from '../../const/const';
+import Logo from '../logo/logo';
+
+const HeaderAuth = (): JSX.Element => {
+  //const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
+  const allOffers = useAppSelector((state) => state.offers);
+  const allFavoritesOffers = allOffers.filter((offer) => offer.isFavorite);
+
+  return(
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <Logo/>
+          <nav className="header__nav">
+            <ul className="header__nav-list">
+              <li className="header__nav-item user">
+                <Link className="header__nav-link header__nav-link--profile" to={`${AppRoute.Favorites}`}>
+                  <div className="header__avatar-wrapper user__avatar-wrapper">
+                  </div>
+                  <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                  <span className="header__favorite-count">{allFavoritesOffers.length}</span>
+                </Link>
+              </li>
+              <li className="header__nav-item">
+                <Link className="header__nav-link" to="/login">
+                  <span className="header__signout">Sign out</span>
+                </Link>
+              </li>
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+};
+
+export default HeaderAuth;
