+++ src/components/main-page/main-page.tsx
@@ -1,9 +1,8 @@
-import Logo from '../logo/logo';
-import { Link } from 'react-router-dom';
+
 import Cards from '../cards-list/cards-list';
 import { OfferPreview } from '../../types/offers-preview';
 import { addPluralEnding } from '../../utils/common';
-import { AppRoute} from '../../const/const';
+import { AuthorizationStatus} from '../../const/const';
 import { useState, useMemo } from 'react';
 import Map from '../map/map';
 import CityList from './city-list';
@@ -15,6 +14,8 @@ import Sorting from './sorting';
 import sort from '../../utils/sort';
 import { fetchOffers } from '../store/api-action';
 import Spinner from '../spinner/spinner';
+import HeaderAuth from './header-auth';
+import HeaderNoAuth from './header-noAuth';
 
 
 const MainPage = (): JSX.Element => {
@@ -28,8 +29,8 @@ const MainPage = (): JSX.Element => {
   const activeCity = useAppSelector((state) => state.city);
   const allOffers = useAppSelector((state) => state.offers);
   const activeSort = useAppSelector((state) => state.sort);
+  const authorizatedStatus = useAppSelector((state) => state.authorizationStatus);
   const filteredOffers = allOffers.filter((offer) => offer.city.name === activeCity);
-  const allFavoritesOffers = allOffers.filter((offer) => offer.isFavorite);
   const cityInfotmation = Cities.find((city) => city.name === activeCity) || Cities[0];
 
   const [activeCard, setActiveCard] = useState<OfferPreview['id'] | null>(null);
@@ -43,30 +44,9 @@ const MainPage = (): JSX.Element => {
 
   return(
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <Logo/>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to={`${AppRoute.Favorites}`}>
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">{allFavoritesOffers.length}</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="/login">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      {authorizatedStatus === AuthorizationStatus.Auth
+        ? <HeaderAuth/>
+        : <HeaderNoAuth/>}
 
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
